add instrumentation and improve tests
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/ibm@238267 138bc75d-0d04-0410-961f-82ee72b054a4

+++ rfc-2464-bounce2/gcc/testsuite/gcc.target/powerpc/bfp/scalar-insert-exp-0.c
@@ -8,10 +8,10 @@
 /* This test should succeed only 64-bit configuration.  */
 #include <altivec.h>
 
-double insert_exponent (unsigned int *exponent_p, 
+double insert_exponent (unsigned long long int *exponent_p, 
 			unsigned long long int *significand_p)
 {
-  unsigned int exponent = *exponent_p;
+  unsigned long long int exponent = *exponent_p;
   unsigned long long int significand = *significand_p;
 
   return scalar_insert_exp (exponent, significand);
